Update to new LearnBase.jl
